ENH: Create new view to embed DICOM browser in main widget
The DICOM browser layout is registered as a single view which spans the whole widget.
Navigating to the DICOM browser module will change to this new layout and hide the data probe, while navigating to a different widget, loading a series (with browser persistence off), or clicking the new "Close" button, will restore the previous layout and settings.